+++ src/realm/views.cpp
@@ -23,108 +23,210 @@
 
 using namespace realm;
 
-LinkChain::LinkChain(size_t single_index)
-    : m_column_indices{single_index}
-{}
+namespace {
+struct IndexPair {
+    size_t index_in_column;
+    size_t index_in_view;
+};
+} // anonymous namespace
 
-LinkChain::LinkChain(std::vector<size_t> chain)
-    : m_column_indices(std::move(chain))
+SortDescriptor::SortDescriptor(Table const& table, std::vector<std::vector<size_t>> column_indices, std::vector<bool> ascending)
+: m_ascending(std::move(ascending))
 {
-    REALM_ASSERT(m_column_indices.size() >= 1);
-}
+    REALM_ASSERT(!column_indices.empty());
+    REALM_ASSERT_EX(m_ascending.empty() || m_ascending.size() == column_indices.size(),
+                    m_ascending.size(), column_indices.size());
+    if (m_ascending.empty())
+        m_ascending.resize(column_indices.size(), true);
+    if (table.is_degenerate())
+        return;
 
-const ColumnBase& LinkChain::init(const ColumnBase* cb, IntegerColumn* row_indexes)
-{
-    REALM_ASSERT(cb != nullptr);
-    REALM_ASSERT(row_indexes != nullptr);
-
-    typedef _impl::TableFriend tf;
-
-    if (m_column_indices.size() > 1) {
-        size_t num_rows = row_indexes->size();
-
-        m_link_translator = std::make_shared<NullableVector>(num_rows);
-
-        std::vector<const LinkColumn*> link_cols;
-        const Table* linked_table = nullptr;
-        const ColumnBase* next_col = cb;
-        for (size_t link_ndx = 0; link_ndx < m_column_indices.size() - 1; link_ndx++) {
-            const LinkColumn* link_col = dynamic_cast<const LinkColumn*>(next_col);
-            // Only last column in link chain is allowed to be non-link
-            if (!link_col) {
+    using tf = _impl::TableFriend;
+    m_columns.resize(column_indices.size());
+    for (size_t i = 0; i < m_columns.size(); ++i) {
+        auto& columns = m_columns[i];
+        auto& indices = column_indices[i];
+        columns.reserve(indices.size());
+        const Table* cur_table = &table;
+        for (auto index : indices) {
+            auto& col = tf::get_column(*cur_table, index);
+            columns.push_back(&col);
+            if (auto link_col = dynamic_cast<const LinkColumn*>(&col)) {
+                cur_table = &link_col->get_target_table();
+            }
+            else if (columns.size() != indices.size()) {
+                // Only last column in link chain is allowed to be non-link
                 throw LogicError(LogicError::type_mismatch);
             }
-            link_cols.push_back(link_col);
-            linked_table = &link_col->get_target_table();
-            next_col = &tf::get_column(*linked_table, m_column_indices[link_ndx + 1]);
         }
+    }
+}
+
+std::vector<std::vector<size_t>> SortDescriptor::get_column_indices() const
+{
+    std::vector<std::vector<size_t>> ret;
+    ret.reserve(m_columns.size());
+    for (auto& cols : m_columns) {
+        std::vector<size_t> indices;
+        indices.reserve(cols.size());
+        for (const ColumnBase* col : cols)
+            indices.push_back(col->get_column_index());
+        ret.push_back(std::move(indices));
+    }
+    return ret;
+}
+
+
+void SortDescriptor::generate_patch(SortDescriptor const& desc, HandoverPatch& patch)
+{
+    if (desc)
+        patch.reset(new SortDescriptorHandoverPatch{desc.get_column_indices(), desc.get_ascending()});
+}
+
+SortDescriptor SortDescriptor::create_from_and_consume_patch(HandoverPatch& patch, Table const& table)
+{
+    SortDescriptor ret;
+    if (patch) {
+        ret = SortDescriptor(table, std::move(patch->columns), std::move(patch->ascending));
+        patch.reset();
+    }
+    return ret;
+}
+
+class SortDescriptor::Sorter {
+public:
+    Sorter(std::vector<std::vector<const ColumnBase*>> const& columns,
+           std::vector<bool> const& ascending, IntegerColumn const& row_indexes);
+
+    bool operator()(IndexPair i, IndexPair j, bool total_ordering = true) const;
+
+private:
+    struct SortColumn {
+        std::vector<bool> is_null;
+        std::vector<size_t> translated_row;
+        const ColumnBase* column;
+        bool ascending;
+    };
+    std::vector<SortColumn> m_columns;
+};
+
+SortDescriptor::Sorter::Sorter(std::vector<std::vector<const ColumnBase*>> const& columns,
+                               std::vector<bool> const& ascending, IntegerColumn const& row_indexes)
+{
+    size_t num_rows = row_indexes.size();
+
+    m_columns.reserve(columns.size());
+    for (size_t i = 0; i < columns.size(); ++i) {
+        m_columns.push_back({{}, {}, columns[i].back(), ascending[i]});
+        if (columns[i].size() == 1) { // no link chain
+            continue;
+        }
+
+        auto& translated_rows = m_columns.back().translated_row;
+        auto& is_null = m_columns.back().is_null;
+        translated_rows.resize(num_rows);
+        is_null.resize(num_rows);
 
         for (size_t row_ndx = 0; row_ndx < num_rows; row_ndx++) {
-            size_t translated_index = row_indexes->get(row_ndx);
-            bool set_null = false;
-            for (const LinkColumn* link_col : link_cols) {
+            size_t translated_index = row_indexes.get(row_ndx);
+            for (size_t j = 0; j + 1 < columns[i].size(); ++j) {
+                auto link_col = static_cast<const LinkColumn*>(columns[i][j]); // type was checked when creating the SortDescriptor
                 if (link_col->is_null(translated_index)) {
-                    set_null = true;
+                    is_null[row_ndx] = true;
                     break;
                 }
-                else {
-                    translated_index = link_col->get_link(translated_index);
-                }
+                translated_index = link_col->get_link(translated_index);
             }
-            REALM_ASSERT_EX(row_ndx < m_link_translator->size(), row_ndx, m_link_translator->size());
-            if (set_null) {
-                (*m_link_translator)[row_ndx] = {};
-            }
-            else {
-                (*m_link_translator)[row_ndx] = { translated_index };
-            }
+            translated_rows[row_ndx] = translated_index;
         }
-        REALM_ASSERT(linked_table);
-        const ColumnBase& last_col_in_chain = tf::get_column(*linked_table, m_column_indices.back());
-        return last_col_in_chain;
     }
-    return *cb; // no link chain, return original column
 }
 
-util::Optional<size_t> LinkChain::translate(size_t index) const
+SortDescriptor::Sorter SortDescriptor::sorter(IntegerColumn const& row_indexes) const
 {
-    util::Optional<size_t> value(index);
-    if (m_link_translator) {
-        REALM_ASSERT_EX(index < m_link_translator->size(), index, m_link_translator->size());
-        value = (*m_link_translator)[index];
-    }
-    return value;
+    return Sorter(m_columns, m_ascending, row_indexes);
 }
 
-
-// Re-sort view according to last used criterias
-void RowIndexes::sort(Sorter& sorting_predicate)
+bool SortDescriptor::Sorter::operator()(IndexPair i, IndexPair j, bool total_ordering) const
 {
+    for (size_t t = 0; t < m_columns.size(); t++) {
+        size_t index_i = i.index_in_column;
+        size_t index_j = j.index_in_column;
+
+        if (!m_columns[t].translated_row.empty()) {
+            bool null_i = m_columns[t].is_null[i.index_in_view];
+            bool null_j = m_columns[t].is_null[j.index_in_view];
+
+            if (null_i && null_j) {
+                continue;
+            }
+            if (null_i || null_j) {
+                // Sort null links at the end if m_ascending[t], else at beginning.
+                return m_columns[t].ascending != null_i;
+            }
+
+            index_i = m_columns[t].translated_row[i.index_in_view];
+            index_j = m_columns[t].translated_row[j.index_in_view];
+        }
+
+        if (int c = m_columns[t].column->compare_values(index_i, index_j))
+            return m_columns[t].ascending ? c > 0 : c < 0;
+    }
+    // make sort stable by using original index as final comparison
+    return total_ordering ? i.index_in_view < j.index_in_view : 0;
+}
+
+void RowIndexes::do_sort(const SortDescriptor& order, const SortDescriptor& distinct)
+{
+    if (!order && !distinct)
+        return;
     size_t sz = size();
     if (sz == 0)
         return;
 
+    // Gather the current rows into a container we can use std algorithms on
+    size_t detached_ref_count = 0;
     std::vector<IndexPair> v;
     v.reserve(sz);
     // always put any detached refs at the end of the sort
     // FIXME: reconsider if this is the right thing to do
     // FIXME: consider specialized implementations in derived classes
     // (handling detached refs is not required in linkviews)
-    size_t detached_ref_count = 0;
     for (size_t t = 0; t < sz; t++) {
         int64_t ndx = m_row_indexes.get(t);
         if (ndx != detached_ref) {
-            v.push_back(IndexPair(ndx, t));
+            v.push_back(IndexPair{static_cast<size_t>(ndx), t});
         }
         else
             ++detached_ref_count;
     }
-    sorting_predicate.init(this);
-    std::stable_sort(v.begin(), v.end(), sorting_predicate);
-    sorting_predicate.cleanup();
+
+    if (distinct) {
+        // Sort by the columns to distinct on
+        auto sorting_predicate = distinct.sorter(m_row_indexes);
+        std::sort(v.begin(), v.end(), std::ref(sorting_predicate));
+
+        // Remove all duplicates
+        v.erase(std::unique(v.begin(), v.end(), [&](auto&& a, auto&& b) {
+            // "not less than" is "equal" since they're sorted
+            return !sorting_predicate(a, b, false);
+        }), v.end());
+
+        // Restore the original order unless we're just going to sort it again anyway
+        if (!order) {
+            std::sort(v.begin(), v.end(), [](auto a, auto b) { return a.index_in_view < b.index_in_view; });
+        }
+    }
+
+    if (order) {
+        auto sorting_predicate = order.sorter(m_row_indexes);
+        std::sort(v.begin(), v.end(), std::ref(sorting_predicate));
+    }
+
+    // Apply the results
     m_row_indexes.clear();
-    for (size_t t = 0; t < sz - detached_ref_count; t++)
-        m_row_indexes.add(v[t].index_in_column);
+    for (auto& pair : v)
+        m_row_indexes.add(pair.index_in_column);
     for (size_t t = 0; t < detached_ref_count; ++t)
         m_row_indexes.add(-1);
 }
@@ -132,26 +234,25 @@ void RowIndexes::sort(Sorter& sorting_predicate)
 // FIXME: this only works (and is only used) for row indexes with memory
 // managed by the default allocator, e.q. for TableViews.
 RowIndexes::RowIndexes(const RowIndexes& source, ConstSourcePayload mode)
-    : m_row_indexes()
-{
 #ifdef REALM_COOKIE_CHECK
-    cookie = source.cookie;
+    : cookie(source.cookie)
 #endif
-    if (mode == ConstSourcePayload::Copy) {
-        if (source.m_row_indexes.is_attached()) {
-            // we only clone if there is something to clone:
-            //m_row_indexes.destroy();
-            // MemRef mem = root->clone_deep(Allocator::get_default());
-            MemRef mem = source.m_row_indexes.clone_deep(Allocator::get_default());
-            m_row_indexes.destroy();
-            m_row_indexes.init_from_mem(Allocator::get_default(), mem);
-        }
+{
+    REALM_ASSERT(&source.m_row_indexes.get_alloc() == &Allocator::get_default());
+
+    if (mode == ConstSourcePayload::Copy && source.m_row_indexes.is_attached()) {
+        MemRef mem = source.m_row_indexes.clone_deep(Allocator::get_default());
+        m_row_indexes.init_from_mem(Allocator::get_default(), mem);
     }
 }
 
 RowIndexes::RowIndexes(RowIndexes& source, MutableSourcePayload)
-    : m_row_indexes()
+#ifdef REALM_COOKIE_CHECK
+    : cookie(source.cookie)
+#endif
 {
+    REALM_ASSERT(&source.m_row_indexes.get_alloc() == &Allocator::get_default());
+
     // move the data payload, but make sure to leave the source array intact or
     // attempts to reuse it for a query rerun will crash (or assert, if lucky)
     // There really *has* to be a way where we don't need to first create an empty
@@ -161,8 +262,4 @@ RowIndexes::RowIndexes(RowIndexes& source, MutableSourcePayload)
         m_row_indexes.init_from_mem(Allocator::get_default(), source.m_row_indexes.get_mem());
         source.m_row_indexes.init_from_ref(Allocator::get_default(), IntegerColumn::create(Allocator::get_default()));
     }
-
-#ifdef REALM_COOKIE_CHECK
-    cookie = source.cookie;
-#endif
 }