+++ test/test_link_query_view.cpp
@@ -1063,19 +1063,14 @@ TEST(LinkList_SortLinkView)
     CHECK_EQUAL(tv.get(2).get_index(), 1);
 
     // Test multi-column sorting
-    std::vector<LinkChain> v;
-    std::vector<bool> a;
-    a.push_back(true);
-    a.push_back(true);
+    std::vector<std::vector<size_t>> v;
+    std::vector<bool> a = {true, true};
+    std::vector<bool> a_false = {false, false};
 
-    std::vector<bool> a_false;
-    a_false.push_back(false);
-    a_false.push_back(false);
-
-    v.push_back(4);
-    v.push_back(1);
-    lvr->sort(v, a_false);
-    tv = lvr->get_sorted_view(v, a_false);
+    v.push_back({4});
+    v.push_back({1});
+    lvr->sort(SortDescriptor{lvr->get_target_table(), v, a_false});
+    tv = lvr->get_sorted_view(SortDescriptor{lvr->get_target_table(), v, a_false});
     CHECK_EQUAL(lvr->get(0).get_index(), 0);
     CHECK_EQUAL(lvr->get(1).get_index(), 2);
     CHECK_EQUAL(lvr->get(2).get_index(), 1);
@@ -1083,8 +1078,8 @@ TEST(LinkList_SortLinkView)
     CHECK_EQUAL(tv.get(1).get_index(), 2);
     CHECK_EQUAL(tv.get(2).get_index(), 1);
 
-    lvr->sort(v, a);
-    tv = lvr->get_sorted_view(v, a);
+    lvr->sort(SortDescriptor{lvr->get_target_table(), v, a});
+    tv = lvr->get_sorted_view(SortDescriptor{lvr->get_target_table(), v, a});
     CHECK_EQUAL(lvr->get(0).get_index(), 1);
     CHECK_EQUAL(lvr->get(1).get_index(), 2);
     CHECK_EQUAL(lvr->get(2).get_index(), 0);
@@ -1092,11 +1087,11 @@ TEST(LinkList_SortLinkView)
     CHECK_EQUAL(tv.get(1).get_index(), 2);
     CHECK_EQUAL(tv.get(2).get_index(), 0);
 
-    v.push_back(2);
+    v.push_back({2});
     a.push_back(true);
 
-    lvr->sort(v, a);
-    tv = lvr->get_sorted_view(v, a);
+    lvr->sort(SortDescriptor{lvr->get_target_table(), v, a});
+    tv = lvr->get_sorted_view(SortDescriptor{lvr->get_target_table(), v, a});
     CHECK_EQUAL(lvr->get(0).get_index(), 1);
     CHECK_EQUAL(lvr->get(1).get_index(), 2);
     CHECK_EQUAL(lvr->get(2).get_index(), 0);
@@ -1108,7 +1103,6 @@ TEST(LinkList_SortLinkView)
     tv.sync_if_needed();
     CHECK_EQUAL(tv.get(0).get_index(), 0);
     CHECK_EQUAL(tv.get(1).get_index(), 1);
-
 }